Add files via upload
Problem Description:
The issue revolved around the inability to perform a new city search after returning to the WeatherForm component from WeatherDisplay. Initially, setting buttonActive to false in the WeatherForm component led to a situation where the component wouldn't render again.

Solution:
The solution involved setting buttonActive to false in the middle component, after WeatherForm was displayed and before WeatherDisplay was rendered. This approach allowed for a seamless transition between the components, enabling users to perform a new search for a different city without encountering rendering issues.